feat: update search command with fusejs
