+++ public/search/search-index.json
@@ -2,47 +2,80 @@
   "posts": [
     {
       "title": "VSCode で日本語がズレる問題を Hack Nerd Font と Cica で解決する",
-      "slug": "vscode-font-settings"
+      "slug": "vscode-font-settings",
+      "description": "VSCode で日本語テキストを編集するときの「なんかもっさりしてるな〜」を、フォント設定を Hack Nerd Font と Cica の組み合わせに変えたら解決した話です。",
+      "tags": [
+        "vscode"
+      ]
     },
     {
       "title": "ミニマリズムを追求した Next.js 15 + MDX ブログテンプレート \"next-minimal-blog\" を作ってみた",
-      "slug": "next-minimal-blog"
+      "slug": "next-minimal-blog",
+      "description": "「ミニマリストアプローチ！公式の Next.js / Vercel 製だけでどこまでできる！？」をテーマに作成した Next.js テンプレート next-minimal-blog を紹介します。",
+      "tags": [
+        "nextjs",
+        "typescript",
+        "tailwindcss"
+      ]
     },
     {
       "title": "Next.js 15 対応とあわせて、ESLint と Prettier を Flat Config に移行した",
-      "slug": "eslint-prettier-flat-config"
+      "slug": "eslint-prettier-flat-config",
+      "description": "Next.js のバージョンアップ (14 -> 15) とあわせて、ESLint と Prettier の設定を Flat Config に移行したメモです。",
+      "tags": [
+        "eslint",
+        "typescript",
+        "vscode"
+      ]
     },
     {
       "title": "npm / yarn / pnpm コマンド違い早見表",
-      "slug": "npm-vs-yarn-vs-pnpm-commands"
+      "slug": "npm-vs-yarn-vs-pnpm-commands",
+      "description": "pnpm をメインで利用するようになってから、パッケージ操作コマンドで npm との違いをググることが増えました。npm, yarn, pnpm のコマンドを早見表としてまとめます。",
+      "tags": [
+        "typescript"
+      ]
     },
     {
       "title": "WordPress でプラグインなしで目次を自動生成する方法",
-      "slug": "wordpress-toc-no-plugin"
+      "slug": "wordpress-toc-no-plugin",
+      "description": "WordPress テーマをカスタマイズして、プラグインなしで記事に目次を追加する手順を紹介します。",
+      "tags": [
+        "wordpress"
+      ]
     },
     {
       "title": "Thunderbird で自動改行の設定が反映されないときの対処法",
-      "slug": "thunderbird-line-break-settings"
+      "slug": "thunderbird-line-break-settings",
+      "description": "Thunderbird でテキストメールを利用している人向けに，送信するときに自動で改行する設定方法を紹介します．",
+      "tags": []
     }
   ],
   "tags": {
     "vscode": {
-      "name": "VSCode"
+      "name": "VSCode",
+      "count": 2
     },
     "nextjs": {
-      "name": "Next.js"
+      "name": "Next.js",
+      "count": 1
     },
     "typescript": {
-      "name": "TypeScript"
+      "name": "TypeScript",
+      "count": 3
     },
     "tailwindcss": {
-      "name": "Tailwind CSS"
+      "name": "Tailwind CSS",
+      "count": 1
     },
     "eslint": {
-      "name": "ESLint"
+      "name": "ESLint",
+      "count": 1
     },
     "wordpress": {
-      "name": "WordPress"
+      "name": "WordPress",
+      "count": 1
     }
-  }
+  },
+  "lastUpdated": "2025-03-23T11:36:59.171Z"
 }
