+++ scripts/generate-search-index.ts
@@ -17,13 +17,18 @@ const OUTPUT_PATH = path.join(
   const posts = allPosts.map((post) => ({
     title: post.metadata.title,
     slug: post.slug,
+    description: post.metadata.description || '',
+    tags: post.metadata.tags || [],
   }));
 
   const usedTags = allPosts
     .flatMap((post) => post.metadata.tags || [])
-    .reduce<Record<string, { name: string }>>((acc, tagSlug) => {
-      if (tags[tagSlug] && !acc[tagSlug]) {
-        acc[tagSlug] = { name: tags[tagSlug].name };
+    .reduce<Record<string, { name: string; count: number }>>((acc, tagSlug) => {
+      if (tags[tagSlug]) {
+        if (!acc[tagSlug]) {
+          acc[tagSlug] = { name: tags[tagSlug].name, count: 0 };
+        }
+        acc[tagSlug].count++;
       }
       return acc;
     }, {});
@@ -31,9 +36,10 @@ const OUTPUT_PATH = path.join(
   const indexData = {
     posts,
     tags: usedTags,
+    lastUpdated: new Date().toISOString(),
   };
 
   fs.writeFileSync(OUTPUT_PATH, JSON.stringify(indexData, null, 2), 'utf-8');
 
-  console.log(`✅ search-index.json を生成しました: ${OUTPUT_PATH}`);
+  console.log(`✅ search index generated successfully: ${OUTPUT_PATH}`);
 })();
